+++ CHANGELOG.md
@@ -7,6 +7,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Added
+
+* `attr::none()` for conditional rendering of an attribute
+* implement `Default` for `Element` and `Attribute`
+
 
 ## [1.4.0] - 2024-12-13